GRPC verify token service done
