+++ Auth-Frontstore/internal/core/ports/server.go
@@ -2,4 +2,5 @@ package ports
 
 type IServer interface {
 	Initialize()
+	InitializeGrpcServer()
 }
